+++ RELEASE-NOTES.md
@@ -21,6 +21,7 @@
 * #587 Avro loader: support for complex types
 * #589 ECharts: YAxis.scale(..) property
 * #590 ECharts: VisualMap.outOfRange
+* #591 ECharts: CalendarModel.yearLabel
 
 ## Release 2.0.0-M5